Refactor the bot with cogs
Move all of the bot slash commands out of the main function (bot.py) and into 'cogs', defined in subfolder 'command_cogs'.
So now, bot.py only has the main bot class and the code to initialize and run the bot in main.

The 'cogs' are loaded in the bot class of FZDBot via 'load_extension'. This allows for easily adding new commands in the future.
Each 'cog' class has its own database connection.

Other changes in this commit:

- Modified error handling in editScore and deleteScore, to catch any Exceptions
- Added error message in "/show" - just in case an event is defined but has no results entered yet
- Changed ISOLATION LEVEL in connect_to_database() to READ COMMITTED, so that each db connection can see the committed changes from other connections